fix: sanitize error messages in log calls to resolve remaining CodeQL alerts
Sanitize err.Error() output in log statements where orchestrator errors
may contain user-derived content (instance names embedded in K8s resource
names). Explicit uint cast in keys.go to satisfy CodeQL integer taint tracking.

+++ control-plane/internal/llmgateway/keys.go
@@ -39,7 +39,7 @@ func EnsureKeysForInstance(instanceID uint, enabledProviderIDs []uint) error {
 		if err := database.DB.Create(&key).Error; err != nil {
 			return fmt.Errorf("create gateway key for instance %d, provider %d: %w", instanceID, providerID, err)
 		}
-		log.Printf("LLM gateway: created key for instance %d, provider %d", instanceID, providerID)
+		log.Printf("LLM gateway: created key for instance %d, provider %d", uint(instanceID), uint(providerID))
 	}
 
 	// Remove keys for disabled providers
